Moved chord parsing to KeyParser
